Factored out Hydra::AdminPolicyBehavior module from Hydra::AdminPolicy class to allow for more flexible implementations of policy objects.
